Adjusting to Google-style II
