+++ src/main/java/mops/portfolios/HardMock.java
@@ -2,8 +2,8 @@ package mops.portfolios;
 
 import mops.portfolios.domain.entry.Entry;
 import mops.portfolios.domain.entry.EntryField;
-import mops.portfolios.domain.Portfolio.Portfolio;
-import mops.portfolios.domain.UserGroup.User;
+import mops.portfolios.domain.portfolio.Portfolio;
+import mops.portfolios.domain.userGroup.User;
 import org.springframework.stereotype.Service;
 
 import java.util.Arrays;
