+++ src/main/java/mops/portfolios/UserSecurity.java
@@ -1,6 +1,6 @@
 package mops.portfolios;
 
-import mops.portfolios.domain.Portfolio.Portfolio;
+import mops.portfolios.domain.portfolio.Portfolio;
 import org.springframework.stereotype.Service;
 
 @Service
