--- src/main/java/mops/portfolios/domain/Portfolio/PortfolioConfig.java
@@ -1,8 +1,0 @@
- package mops.portfolios.domain.Portfolio;
-
- import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
- import org.springframework.transaction.annotation.EnableTransactionManagement;
-
- @EnableJpaRepositories
- @EnableTransactionManagement
- class Config{}
