+++ src/main/java/mops/portfolios/domain/portfolio/PortfolioConfig.java
@@ -1,0 +1,8 @@
+package mops.portfolios.domain.portfolio;
+
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+import org.springframework.transaction.annotation.EnableTransactionManagement;
+
+@EnableJpaRepositories
+@EnableTransactionManagement
+class Config{}
