+++ src/main/java/mops/portfolios/domain/portfolio/PortfolioRepository.java
@@ -1,13 +1,12 @@
-package mops.portfolios.domain.Portfolio;
+package mops.portfolios.domain.portfolio;
 
 import java.util.List;
-
 import org.springframework.data.repository.CrudRepository;
 
 public interface PortfolioRepository extends CrudRepository<Portfolio, Long> {
-    
-    List<Portfolio> findAllByUserId(String userId);
 
-    List<Portfolio> findAllByGroupId(Long groupId);
+  List<Portfolio> findAllByUserId(String userId);
+
+  List<Portfolio> findAllByGroupId(Long groupId);
 
 }
