+++ src/main/java/mops/portfolios/domain/userGroup/UserGroupRepository.java
@@ -1,4 +1,4 @@
-package mops.portfolios.domain.UserGroup;
+package mops.portfolios.domain.userGroup;
 
 import org.springframework.data.repository.CrudRepository;
